Add files via upload
Flash_Card_Game is an interactive flashcard application designed to help users learn and memorize foreign language vocabulary. The app displays French words on one side of the flashcard, and after a short delay, flips to reveal the English translation. Users can mark words they know as "known," and these words are saved in a separate list for future review. With a simple and clean user interface built using tkinter, Flash_Card_Game is a great tool for learners looking to expand their vocabulary in a fun and efficient way.